fix: Removed folders
