--- mediasms/models/client.go
@@ -1,8 +1,0 @@
-package models
-
-// Client struct
-type Client struct {
-	Username string
-	Password string
-	Prefix   string
-}
